(broken) minor refactor
