+++ frontend/src/api/entities/ProjectSkillEntity.ts
@@ -42,3 +42,7 @@ export class ProjectSkill {
     name: string;
     project: string;
 }
+
+export function ProjectSkillFromIProjectSkill() {
+
+}
