+++ frontend/src/pages/projects/create.tsx
@@ -1,7 +1,7 @@
 import type { NextPage } from "next";
 import NavBar from "../../components/util/navBar";
 import styles from "../../styles/projects/createProject.module.css";
-import { ProjectForm } from "../../components/project/projectForm";
+import { ProjectForm } from "../../components/project/projectForm/projectForm";
 import { ProjectFormSubmitHandler } from "../../handlers/projectFormSubmitHandler";
 
 const CreateProject: NextPage = () => {
